Split out TransactionContext classes
The classes are static and quite large, splitting them out into separate
files allows for better readability.
